refactor: Move WheelchairAccessibility(Feature|Request) to preferences
