+++ src/main/java/org/opentripplanner/routing/algorithm/raptoradapter/transit/cost/McCostParams.java
@@ -6,7 +6,7 @@ import java.util.function.DoubleFunction;
 import javax.annotation.Nullable;
 import org.opentripplanner.routing.api.request.RequestFunctions;
 import org.opentripplanner.routing.api.request.RouteRequest;
-import org.opentripplanner.routing.api.request.WheelchairAccessibilityRequest;
+import org.opentripplanner.routing.api.request.preference.WheelchairAccessibilityRequest;
 import org.opentripplanner.transit.raptor.api.transit.RaptorTripSchedule;
 import org.opentripplanner.util.lang.ToStringBuilder;
