+++ src/main/java/org/opentripplanner/routing/api/request/preference/WheelchairPreferences.java
@@ -2,7 +2,6 @@ package org.opentripplanner.routing.api.request.preference;
 
 import java.io.Serializable;
 import javax.annotation.Nonnull;
-import org.opentripplanner.routing.api.request.WheelchairAccessibilityRequest;
 
 // TODO VIA: Remove this class and use existing WheelchairAccessibilityRequest instead
 public class WheelchairPreferences implements Cloneable, Serializable {
