+++ src/test/java/org/opentripplanner/routing/api/request/WheelchairAccessibilityRequestTest.java
@@ -1,11 +1,12 @@
 package org.opentripplanner.routing.api.request;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.opentripplanner.routing.api.request.WheelchairAccessibilityFeature.ofOnlyAccessible;
+import static org.opentripplanner.routing.api.request.preference.WheelchairAccessibilityFeature.ofOnlyAccessible;
 
 import java.util.stream.Stream;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
+import org.opentripplanner.routing.api.request.preference.WheelchairAccessibilityRequest;
 import org.opentripplanner.test.support.VariableSource;
 
 class WheelchairAccessibilityRequestTest {
